world 周りをリファクタリングする
